requirements.txt
Django>=4.2,<5
djangorestframework
pillow
django-crispy-forms
crispy-bootstrap5
gunicorn
whitenoise
dj-database-url
psycopg2-binary
python-dotenv
